[test_loop] Cleanup old test loop code
