+++ chain/chain/src/test_utils.rs
@@ -1,5 +1,4 @@
 mod kv_runtime;
-pub mod test_loop;
 mod validator_schedule;
 
 use std::cmp::Ordering;
