+++ chain/network/src/client.rs
@@ -1,6 +1,7 @@
 use crate::network_protocol::StateResponseInfo;
 use crate::types::{NetworkInfo, ReasonForBan};
 use near_async::messaging::AsyncSender;
+use near_async::{MultiSend, MultiSendMessage, MultiSenderFrom};
 use near_primitives::block::{Approval, Block, BlockHeader};
 use near_primitives::challenge::Challenge;
 use near_primitives::errors::InvalidTxError;
@@ -116,9 +117,7 @@ pub struct AnnounceAccountRequest(pub Vec<(AnnounceAccount, Option<EpochId>)>);
 #[rtype(result = "()")]
 pub struct ChunkEndorsementMessage(pub ChunkEndorsement);
 
-#[derive(
-    Clone, near_async::MultiSend, near_async::MultiSenderFrom, near_async::MultiSendMessage,
-)]
+#[derive(Clone, MultiSend, MultiSenderFrom, MultiSendMessage)]
 #[multi_send_message_derive(Debug)]
 #[multi_send_input_derive(Debug, Clone, PartialEq, Eq)]
 pub struct ClientSenderForNetwork {
