--- core/async/src/examples/mod.rs
@@ -1,7 +1,0 @@
-mod actix_component;
-mod actix_component_test;
-mod async_component;
-mod async_component_test;
-mod multi_instance_test;
-mod sum_numbers;
-mod sum_numbers_test;
