+++ core/async/src/lib.rs
@@ -3,8 +3,6 @@ pub use near_async_derive::{MultiSend, MultiSendMessage, MultiSenderFrom};
 pub mod actix;
 pub mod actix_wrapper;
 pub mod break_apart;
-#[cfg(test)]
-mod examples;
 mod functional;
 pub mod futures;
 pub mod messaging;
